requirements.txt
beautifulsoup4==4.13.4
comtypes==1.4.11
geopy==2.4.1
gevent
google-genai==1.29.0
googlesearch-python==1.3.0
groq==0.31.0
keyboard==0.13.5
numpy==2.2.6
openai==1.99.1
opencv-python==4.12.0.88
playwright
PyAudio==0.2.14
PyAutoGUI==0.9.54
piper-tts[http]
pycaw==20240210
pygame==2.6.1
python-dotenv==1.1.1
pywizlight==0.6.3
requests==2.32.4
scipy==1.16.1
sounddevice==0.5.2
soundfile
vosk==0.3.45
ytmusicapi==1.11.3
./playsound-1.3.0-py3-none-any.whl